Move restore_original_embeddings(...) to _shared/ directory.
